requirements.txt.txt
streamlit
openai-whisper
torch
setuptools